back/requirements.txt
dnspython~=2.6.1
idna~=3.8
httpcore~=1.0.5
httpx~=0.26.0
cryptography~=42.0.8
anyio~=4.4.0
h11~=0.14.0
pip~=23.2.1
wheel~=0.41.2
setuptools~=68.2.0
typing_extensions~=4.12.2
exceptiongroup~=1.2.2
sniffio~=1.3.1
click~=8.1.7
certifi~=2024.8.30
libipld~=1.2.3
pydantic~=2.8.2
pydantic_core~=2.20.1
pycparser~=2.22
websockets~=13.0.1
atproto~=0.0.52
python-dotenv~=1.0.1
flask~=2.3.3
requests~=2.31.0